[issue-485] fixing pylint
